backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
django-filter>=24.3
drf-spectacular>=0.27
psycopg[binary]>=3.2
django-environ>=0.11
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
user-agents>=2.2
requests>=2.31
